[Change] DeepDanbooruImageStore 增加目录文件创建; [Fix] DeepDanbooruDatabase 完成putImageMetaData方法; [Update] .gitignore 更新忽略列表; [Add] build.gradle 增加'org.slf4j:slf4j-api'依赖项; [Change] IllustFilter 调整tags参数修饰符; [Add] MetadataDatabase 增加'getTags'方法; [Fix] PagesCountFilter 修复条件错误问题; [Add] PixivDownloader 增加内部方法'getFilter'以允许子类提前筛选作品; [Fix] PixivDownloader 暂时修复ByteArrayInputStream无法重置的bug; [Update] PixivSearchDownloader 完成Pixiv搜索下载器; [Update] MainClass 完成一个临时使用的主类;

+++ src/main/java/net/lamgc/pixiv/downloader/deepdanbooru/DeepDanbooruImageStore.java
@@ -19,8 +19,11 @@ public class DeepDanbooruImageStore implements ImageFileStore {
 
     private final Map<String, File> hashDirMap = new HashMap<>();
 
-    public DeepDanbooruImageStore(File storeDir) {
+    public DeepDanbooruImageStore(File storeDir) throws IOException {
         this.storeDir = storeDir;
+        if(!storeDir.exists() && !storeDir.mkdirs()) {
+            throw new IOException("Directory create failure!");
+        }
     }
 
     @Override
@@ -30,6 +33,9 @@ public class DeepDanbooruImageStore implements ImageFileStore {
         ByteStreams.copy(imageInputStream, bufferStream);
         String md5 = MessageDigestUtils.md5ToString(bufferStream.toByteArray(), true);
         File file = new File(getHashStoreDir(md5), md5 + "." + fileExtName);
+        if(!file.exists() && !file.createNewFile()) {
+            throw new IOException("create file failure");
+        }
         Files.write(file.toPath(), bufferStream.toByteArray(), StandardOpenOption.WRITE);
     }
